@@ -4,8 +4,8 @@
   "main": "index.js",
   "scripts": {
     "typecheck": "tsc",
-    "format": "prettier -w ./*.js ./*.ts packages",
-    "lint": "eslint packages",
+    "format": "prettier -w ./*.ts ./*.mjs packages",
+    "lint": "eslint ./*.ts ./*.mjs packages",
     "publish-devtools": "cd $REPLAY_DIR/devtools && cd packages/shared && npx --yes yalc publish --private && cd ../../packages/protocol && npx yalc publish --private && cd ../../packages/replay-next && npx yalc publish --private",
     "link-devtools": "npx yalc link shared && npx yalc link protocol && npx yalc link replay-next",
     "yalc-devtools": "publish-devtools && link-devtools",
@@ -18,7 +18,7 @@
     "packages/*"
   ],
   "importSort": {
-    ".js, .jsx, .ts, .tsx": {
+    ".mjs, .js, .jsx, .ts, .tsx": {
       "style": "module",
       "parser": "typescript"
     }
@@ -27,35 +27,35 @@
     "@replayio/protocol": "^0.76.0",
     "cli": "workspace:*",
     "commander": "^12.1.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
     "graphql": "^16.9.0",
     "lodash": "^4.17.21",
-    "uuid": "^10.0.0",
     "ws": "^8.18.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.13.0",
+    "@eslint/js": "^9.15.0",
     "@tsconfig/node22": "^22.0.0",
     "@types/eslint__js": "^8.42.3",
     "@types/jest": "^29.5.14",
-    "@types/lodash": "^4.14.181",
-    "@types/node": "^22.7.7",
+    "@types/lodash": "^4.17.13",
+    "@types/node": "^22.10.0",
     "@types/uuid": "^10.0.0",
-    "@types/ws": "^8.5.12",
-    "eslint": "^9.13.0",
+    "@types/ws": "^8.5.13",
+    "eslint": "^9.15.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.2.1",
     "import-sort-style-module": "^6.0.0",
     "jest": "^29.7.0",
     "jest-extended": "^4.0.2",
     "jsonc-parser": "^3.3.1",
-    "prettier": "^3.3.3",
+    "prettier": "^3.4.1",
     "prettier-plugin-import-sort": "^0.0.7",
     "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
     "tsconfig-paths": "^4.2.0",
     "tsconfig-paths-jest": "^0.0.1",
-    "tsx": "^4.19.1",
-    "typescript": "^5.6.3",
-    "typescript-eslint": "^8.11.0"
+    "tsx": "^4.19.2",
+    "typescript": "^5.7.2",
+    "typescript-eslint": "^8.16.0"
   },
-  "packageManager": "yarn@4.5.1"
+  "packageManager": "yarn@4.5.3"
 }
