@@ -18,8 +18,7 @@ const SymptomPoint: ExecutionPoint = "78858008544042601258383216576823298";
 
   const statement = await point.queryStatement();
   const stack = await point.queryRichStack();
-  
+
   console.log("Statement at point:", statement);
   console.log("Stack at point:", stack);
-  
 })();
